Rename morpheusvm controller to vm
