+++ examples/morpheusvm/cmd/morpheus-cli/cmd/handler.go
@@ -18,7 +18,7 @@ import (
 	"github.com/ava-labs/hypersdk/crypto/ed25519"
 	"github.com/ava-labs/hypersdk/crypto/secp256r1"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
-	"github.com/ava-labs/hypersdk/examples/morpheusvm/controller"
+	"github.com/ava-labs/hypersdk/examples/morpheusvm/vm"
 	"github.com/ava-labs/hypersdk/pubsub"
 	"github.com/ava-labs/hypersdk/utils"
 )
@@ -39,7 +39,7 @@ func (h *Handler) Root() *cli.Handler {
 
 func (h *Handler) DefaultActor() (
 	ids.ID, *cli.PrivateKey, chain.AuthFactory,
-	*jsonrpc.JSONRPCClient, *controller.JSONRPCClient, *ws.WebSocketClient, error,
+	*jsonrpc.JSONRPCClient, *vm.JSONRPCClient, *ws.WebSocketClient, error,
 ) {
 	addr, priv, err := h.h.GetDefaultKey(true)
 	if err != nil {
@@ -77,14 +77,14 @@ func (h *Handler) DefaultActor() (
 			Address: addr,
 			Bytes:   priv,
 		}, factory, jcli,
-		controller.NewJSONRPCClient(
+		vm.NewJSONRPCClient(
 			uris[0],
 		), ws, nil
 }
 
 func (*Handler) GetBalance(
 	ctx context.Context,
-	cli *controller.JSONRPCClient,
+	cli *vm.JSONRPCClient,
 	addr codec.Address,
 ) (uint64, error) {
 	saddr, err := codec.AddressBech32(consts.HRP, addr)
@@ -138,7 +138,7 @@ func (*Controller) ParseAddress(addr string) (codec.Address, error) {
 }
 
 func (*Controller) GetParser(uri string) (chain.Parser, error) {
-	cli := controller.NewJSONRPCClient(uri)
+	cli := vm.NewJSONRPCClient(uri)
 	return cli.Parser(context.TODO())
 }
 
@@ -147,7 +147,7 @@ func (*Controller) HandleTx(tx *chain.Transaction, result *chain.Result) {
 }
 
 func (*Controller) LookupBalance(address string, uri string) (uint64, error) {
-	cli := controller.NewJSONRPCClient(uri)
+	cli := vm.NewJSONRPCClient(uri)
 	balance, err := cli.Balance(context.TODO(), address)
 	return balance, err
 }
