+++ examples/morpheusvm/cmd/morpheus-cli/cmd/spam.go
@@ -18,14 +18,14 @@ import (
 	"github.com/ava-labs/hypersdk/crypto/secp256r1"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/actions"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
-	"github.com/ava-labs/hypersdk/examples/morpheusvm/controller"
+	"github.com/ava-labs/hypersdk/examples/morpheusvm/vm"
 	"github.com/ava-labs/hypersdk/pubsub"
 	"github.com/ava-labs/hypersdk/utils"
 )
 
 type SpamHelper struct {
 	keyType string
-	cli     *controller.JSONRPCClient
+	cli     *vm.JSONRPCClient
 	ws      *ws.WebSocketClient
 }
 
@@ -51,7 +51,7 @@ func (*SpamHelper) GetFactory(pk *cli.PrivateKey) (chain.AuthFactory, error) {
 }
 
 func (sh *SpamHelper) CreateClient(uri string) error {
-	sh.cli = controller.NewJSONRPCClient(uri)
+	sh.cli = vm.NewJSONRPCClient(uri)
 	ws, err := ws.NewWebSocketClient(uri, ws.DefaultHandshakeTimeout, pubsub.MaxPendingMessages, pubsub.MaxReadMessageSize)
 	if err != nil {
 		return err
